% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-details.R
\name{get_details}
\alias{get_details}
\title{Retrieve details (document links) from a jurisprudence record entry}
\usage{
get_details(detail_id_or_url)
}
\arguments{
\item{detail_id_or_url}{either a number (e.g. \code{2606}) or a full URL
(i.e. the \code{detail_url} from a \code{\link[=juris_search]{juris_search()}} query result)}
}
\value{
data frame of document links
}
\description{
Retrieve details (document links) from a jurisprudence record entry
}
\examples{
xdf <- juris_search(year_start = 2019, year_end = 2020)
get_details(xdf$detail_url[1])
get_details(2606)
}
