% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juris-search.R
\name{juris_search}
\alias{juris_search}
\title{Search the U.N. Jurisprudence Database}
\usage{
juris_search(
  keyword = "",
  search_type = c("all", "any"),
  year_start = format(Sys.Date(), "\%Y"),
  year_end = format(Sys.Date(), "\%Y"),
  symbol = "",
  communication_number = "",
  session = "",
  decision_type = c("all", "admissibility", "adoption", "discontinuance",
    "inadmissibility", "opinion", "revised")
)
}
\arguments{
\item{keyword}{keyword}

\item{search_type}{all or any}

\item{year_start, year_end}{year start/end}

\item{symbol}{symbol}

\item{communication_number}{comm #}

\item{session}{session}

\item{decision_type}{decision type}
}
\description{
Search the U.N. Jurisprudence Database
}
\examples{
juris_search(year_start = 2019, year_end = 2020)
}
\references{
\url{https://juris.ohchr.org/search/documents}
}
