% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unjuris-package.R
\docType{package}
\name{unjuris}
\alias{unjuris}
\alias{unjuris-package}
\title{Search the U.N. Jurisprudence Database}
\description{
The jurisprudence database is intended to be a single source of the human rights recommendations
and findings issued by all above committees in their work on individual cases. It enables the general
public, governments, civil society organizations, United Nations partners and international regional
mechanisms to research the vast body of legal interpretation of international human rights law as
it has evolved over the past years. Tools are provided to search and retrieve search results from
the datbase.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/unjuris}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/unjuris/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
